Remove unused stuff, point to new stuff
