+++ src/geometry/mod.rs
@@ -1,42 +1,10 @@
 //! Geometry module.
 
 mod angle;
-mod point;
 mod real;
-mod size;
 
 pub(crate) use angle::angle_consts;
 pub(crate) use angle::Trigonometry;
 pub use angle::{Angle, AngleUnit};
-pub use point::Point;
+pub use embedded_graphics_core::geometry::{Dimensions, OriginDimensions, Point, Size};
 pub(crate) use real::Real;
-pub use size::Size;
-
-use crate::primitives::Rectangle;
-
-/// Adds the ability to get the dimensions/position of a graphics object
-///
-/// This **should** be implemented for all builtin embedded-graphics primitives and fonts. Third party
-/// implementations do not have to implement this trait as an object may not have a known size. If
-/// the object _does_ have a known size, this trait **should** be implemented.
-pub trait Dimensions {
-    /// Returns the bounding box.
-    fn bounding_box(&self) -> Rectangle;
-}
-
-/// Dimensions with `top_left` of the bounding box at `(0, 0)`.
-///
-/// A blanket implementation of `Dimensions` is provided for all types that implement this trait.
-pub trait OriginDimensions {
-    /// Returns the size of the bounding box.
-    fn size(&self) -> Size;
-}
-
-impl<T> Dimensions for T
-where
-    T: OriginDimensions,
-{
-    fn bounding_box(&self) -> Rectangle {
-        Rectangle::new(Point::zero(), self.size())
-    }
-}
