+++ src/primitives/rectangle/mod.rs
@@ -1,67 +1,21 @@
 //! The rectangle primitive. Also good for drawing squares.
 
-mod points;
 mod styled;
 
 use crate::{
-    geometry::{Dimensions, Point, Size},
+    geometry::{Point, Size},
     primitives::{ContainsPoint, OffsetOutline, Primitive},
     transform::Transform,
-    SaturatingCast,
 };
-use core::{
-    cmp::min,
-    ops::{Range, RangeInclusive},
-};
-pub use points::Points;
+use core::ops::RangeInclusive;
+pub use embedded_graphics_core::rectangle::{AnchorPoint, Points, Rectangle};
 pub use styled::StyledPixels;
-
-/// Rectangle primitive
-///
-/// # Examples
-///
-/// ## Create some rectangles with different styles
-///
-/// ```rust
-/// use embedded_graphics::{
-///     pixelcolor::Rgb565, prelude::*, primitives::Rectangle, style::PrimitiveStyleBuilder,
-/// };
-/// # use embedded_graphics::mock_display::MockDisplay;
-/// # let mut display = MockDisplay::default();
-///
-/// // Rectangle with red 3 pixel wide stroke and green fill with the top left corner at (30, 20) and
-/// // a size of (10, 15)
-/// let style = PrimitiveStyleBuilder::new()
-///     .stroke_color(Rgb565::RED)
-///     .stroke_width(3)
-///     .fill_color(Rgb565::GREEN)
-///     .build();
-///
-/// Rectangle::new(Point::new(30, 20), Size::new(10, 15))
-///     .into_styled(style)
-///     .draw(&mut display)?;
-///
-/// // Rectangle with translation applied
-/// Rectangle::new(Point::new(30, 20), Size::new(10, 15))
-///     .translate(Point::new(-20, -10))
-///     .into_styled(style)
-///     .draw(&mut display)?;
-/// # Ok::<(), core::convert::Infallible>(())
-/// ```
-#[derive(Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Default)]
-pub struct Rectangle {
-    /// Top left point of the rectangle.
-    pub top_left: Point,
-
-    /// Size of the rectangle.
-    pub size: Size,
-}
 
 impl Primitive for Rectangle {
     type PointsIter = Points;
 
     fn points(&self) -> Self::PointsIter {
-        Points::new(self)
+        self.points()
     }
 }
 
@@ -74,335 +28,6 @@ impl ContainsPoint for Rectangle {
         } else {
             false
         }
-    }
-}
-
-impl Dimensions for Rectangle {
-    fn bounding_box(&self) -> Rectangle {
-        *self
-    }
-}
-
-impl Rectangle {
-    /// Creates a new rectangle from the top left point and the size.
-    pub const fn new(top_left: Point, size: Size) -> Self {
-        Rectangle { top_left, size }
-    }
-
-    /// Creates a new rectangle from two corners.
-    pub fn with_corners(corner_1: Point, corner_2: Point) -> Self {
-        let left = min(corner_1.x, corner_2.x);
-        let top = min(corner_1.y, corner_2.y);
-
-        Rectangle {
-            top_left: Point::new(left, top),
-            size: Size::from_bounding_box(corner_1, corner_2),
-        }
-    }
-
-    /// Creates a new rectangle from the center point and the size.
-    ///
-    /// For rectangles with even width and/or height the top left corner doesn't
-    /// align with the pixel grid. Because of this the coordinates of the top left
-    /// corner will be rounded up to the nearest integer coordinate.
-    pub fn with_center(center: Point, size: Size) -> Self {
-        Rectangle {
-            top_left: center - size.center_offset(),
-            size,
-        }
-    }
-
-    /// Returns a zero sized rectangle.
-    pub(crate) fn zero() -> Rectangle {
-        Rectangle::new(Point::zero(), Size::zero())
-    }
-
-    /// Returns the center of this rectangle.
-    ///
-    /// For rectangles with even width and/or height the returned value is rounded down
-    /// to the nearest integer coordinate.
-    pub fn center(&self) -> Point {
-        self.top_left + self.size.center_offset()
-    }
-
-    /// Returns the bottom right corner of this rectangle.
-    ///
-    /// Because the smallest rectangle that can be represented by its corners
-    /// has a size of 1 x 1 pixels, this function returns `None` if the width or
-    /// height of the rectangle is zero.
-    pub fn bottom_right(&self) -> Option<Point> {
-        if self.size.width > 0 && self.size.height > 0 {
-            Some(self.top_left + self.size - Point::new(1, 1))
-        } else {
-            None
-        }
-    }
-
-    /// Returns a new `Rectangle` containing the intersection of `self` and `other`.
-    ///
-    /// If no intersection is present, this method will return a zero sized rectangle.
-    ///
-    /// # Examples
-    ///
-    /// ## Intersection
-    ///
-    /// This example draws two rectangles to a mock display using the `.` character, along with
-    /// their intersection shown with `#` characters.
-    ///
-    /// ```rust
-    /// use embedded_graphics::{
-    ///     mock_display::MockDisplay, pixelcolor::BinaryColor, prelude::*, primitives::Rectangle,
-    ///     style::PrimitiveStyle,
-    /// };
-    ///
-    /// let mut display = MockDisplay::new();
-    /// # display.set_allow_overdraw(true);
-    ///
-    /// let rect1 = Rectangle::new(Point::zero(), Size::new(7, 8));
-    /// let rect2 = Rectangle::new(Point::new(2, 3), Size::new(10, 7));
-    ///
-    /// let intersection = rect1.intersection(&rect2);
-    ///
-    /// rect1
-    ///     .into_styled(PrimitiveStyle::with_stroke(BinaryColor::Off, 1))
-    ///     .draw(&mut display)?;
-    ///
-    /// rect2
-    ///     .into_styled(PrimitiveStyle::with_stroke(BinaryColor::Off, 1))
-    ///     .draw(&mut display)?;
-    ///
-    /// intersection
-    ///     .into_styled(PrimitiveStyle::with_stroke(BinaryColor::On, 1))
-    ///     .draw(&mut display)?;
-    ///
-    /// assert_eq!(
-    ///     display,
-    ///     MockDisplay::from_pattern(&[
-    ///         ".......     ",
-    ///         ".     .     ",
-    ///         ".     .     ",
-    ///         ". #####.....",
-    ///         ". #   #    .",
-    ///         ". #   #    .",
-    ///         ". #   #    .",
-    ///         "..#####    .",
-    ///         "  .        .",
-    ///         "  ..........",
-    ///     ])
-    /// );
-    /// # Ok::<(), core::convert::Infallible>(())
-    /// ```
-    ///
-    /// ## No intersection
-    ///
-    /// This example creates two rectangles with no intersection between them. In this case,
-    /// `intersection` returns a zero-sized rectangle.
-    ///
-    /// ```rust
-    /// use embedded_graphics::{prelude::*, primitives::Rectangle, style::PrimitiveStyle};
-    ///
-    /// let rect1 = Rectangle::new(Point::zero(), Size::new(7, 8));
-    /// let rect2 = Rectangle::new(Point::new(10, 15), Size::new(10, 7));
-    ///
-    /// let intersection = rect1.intersection(&rect2);
-    ///
-    /// assert_eq!(intersection.size, Size::zero());
-    /// # Ok::<(), core::convert::Infallible>(())
-    /// ```
-    pub fn intersection(&self, other: &Rectangle) -> Rectangle {
-        match (other.bottom_right(), self.bottom_right()) {
-            (Some(other_bottom_right), Some(self_bottom_right)) => {
-                if overlaps(
-                    self.top_left.x..=self_bottom_right.x,
-                    other.top_left.x..=other_bottom_right.x,
-                ) && overlaps(
-                    self.top_left.y..=self_bottom_right.y,
-                    other.top_left.y..=other_bottom_right.y,
-                ) {
-                    return Rectangle::with_corners(
-                        self.top_left.component_max(other.top_left),
-                        self_bottom_right.component_min(other_bottom_right),
-                    );
-                }
-            }
-            (Some(_other_bottom_right), None) => {
-                // Check if zero sized self is inside other
-                if other.contains(self.top_left) {
-                    return self.clone();
-                }
-            }
-            (None, Some(_self_bottom_right)) => {
-                // Check if zero sized other is inside self
-                if self.contains(other.top_left) {
-                    return other.clone();
-                }
-            }
-            (None, None) => (),
-        };
-
-        // No overlap present
-        Rectangle::zero()
-    }
-
-    /// Returns a resized copy of this rectangle.
-    ///
-    /// The rectangle is resized relative to the given anchor point.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use embedded_graphics::{prelude::*, primitives::rectangle::{Rectangle, AnchorPoint}};
-    ///
-    /// let rect = Rectangle::new(Point::new(20, 20), Size::new(10, 20));
-    /// let resized = rect.resized(Size::new(20, 10), AnchorPoint::Center);
-    ///
-    /// assert_eq!(resized, Rectangle::new(Point::new(15, 25), Size::new(20, 10)));
-    /// ```
-    pub fn resized(&self, size: Size, anchor_point: AnchorPoint) -> Self {
-        // Assume size = 1 for zero sized dimensions.
-        let one = Size::new_equal(1);
-        let delta = Point::zero() + self.size.component_max(one) - size.component_max(one);
-
-        let top_left = self.top_left
-            + match anchor_point {
-                AnchorPoint::TopLeft => Point::zero(),
-                AnchorPoint::TopCenter => delta.x_axis() / 2,
-                AnchorPoint::TopRight => delta.x_axis(),
-                AnchorPoint::CenterLeft => delta.y_axis() / 2,
-                AnchorPoint::Center => delta / 2,
-                AnchorPoint::CenterRight => Point::new(delta.x, delta.y / 2),
-                AnchorPoint::BottomLeft => delta.y_axis(),
-                AnchorPoint::BottomCenter => Point::new(delta.x / 2, delta.y),
-                AnchorPoint::BottomRight => delta,
-            };
-
-        Self::new(top_left, size)
-    }
-
-    /// Returns an anchor point.
-    ///
-    /// # Examples
-    /// ```
-    /// use embedded_graphics::{prelude::*, primitives::rectangle::{Rectangle, AnchorPoint}};
-    ///
-    /// let mut rect = Rectangle::new(Point::new(20, 20), Size::new(11, 21));
-    ///
-    /// assert_eq!(rect.anchor_point(AnchorPoint::TopLeft), Point::new(20, 20));
-    /// assert_eq!(rect.anchor_point(AnchorPoint::BottomCenter), Point::new(25, 40));
-    /// ```
-    pub fn anchor_point(&self, anchor_point: AnchorPoint) -> Point {
-        // Assume size = 1 for zero sized dimensions.
-        let one = Size::new_equal(1);
-        let delta = Point::zero() + self.size.component_max(one) - one;
-
-        self.top_left
-            + match anchor_point {
-                AnchorPoint::TopLeft => Point::zero(),
-                AnchorPoint::TopCenter => delta.x_axis() / 2,
-                AnchorPoint::TopRight => delta.x_axis(),
-                AnchorPoint::CenterLeft => delta.y_axis() / 2,
-                AnchorPoint::Center => delta / 2,
-                AnchorPoint::CenterRight => Point::new(delta.x, delta.y / 2),
-                AnchorPoint::BottomLeft => delta.y_axis(),
-                AnchorPoint::BottomCenter => Point::new(delta.x / 2, delta.y),
-                AnchorPoint::BottomRight => delta,
-            }
-    }
-
-    /// Returns the range of Y coordinates in this rectangle.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use embedded_graphics::{prelude::*, primitives::Rectangle};
-    ///
-    /// let rect = Rectangle::new(Point::new(10, 20), Size::new(3, 4));
-    /// assert_eq!(rect.rows(), 20..24);
-    /// ```
-    ///
-    /// By combining this method with [`columns`] it is possible to iterate over all pixels inside
-    /// the rectangle. This can be more flexible than using the [`points`] iterator, for example,
-    /// if a different iteration order is required or some operations should be called once per row.
-    ///
-    /// ```
-    /// use embedded_graphics::{prelude::*, primitives::Rectangle};
-    ///
-    /// let rect = Rectangle::new(Point::new(10, 20), Size::new(3, 4));
-    ///
-    /// // Iterate over the y coordinates of the rows in reverse order.
-    /// for y in rect.rows().rev() {
-    ///     for x in rect.columns() {
-    ///         // use x, y coordinates
-    ///     }
-    /// }
-    /// ```
-    ///
-    /// [`columns`]: #method.columns
-    /// [`points`]: ../trait.Primitive.html#tymethod.points
-    pub fn rows(&self) -> Range<i32> {
-        self.top_left.y
-            ..self
-                .top_left
-                .y
-                .saturating_add(self.size.height.saturating_cast())
-    }
-
-    /// Returns the range of X coordinates in this rectangle.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use embedded_graphics::{prelude::*, primitives::Rectangle};
-    ///
-    /// let rect = Rectangle::new(Point::new(10, 20), Size::new(3, 4));
-    ///
-    /// assert_eq!(rect.columns(), 10..13);
-    /// ```
-    ///
-    /// By combining this method with [`rows`] it is possible to iterator over all pixels inside
-    /// the rectangle. This can be more flexible than using the [`points`] iterator, for example,
-    /// if a different iteration order is required or some operations should be called once per row.
-    ///
-    /// ```
-    /// use embedded_graphics::{prelude::*, primitives::Rectangle};
-    ///
-    /// let rect = Rectangle::new(Point::new(10, 20), Size::new(3, 4));
-    ///
-    /// // Iterate over all points starting from the top right corner and advancing downwards.
-    /// for x in rect.columns().rev() {
-    ///     for y in rect.rows() {
-    ///         // use x, y coordinates
-    ///     }
-    /// }
-    /// ```
-    ///
-    /// [`rows`]: #method.rows
-    /// [`points`]: ../trait.Primitive.html#tymethod.points
-    pub fn columns(&self) -> Range<i32> {
-        self.top_left.x
-            ..self
-                .top_left
-                .x
-                .saturating_add(self.size.width.saturating_cast())
-    }
-
-    /// Returns `true` is the rectangle is zero sized.
-    ///
-    /// A rectangle is zero sized if the width or height are zero.
-    ///
-    /// # Examples
-    /// ```
-    /// use embedded_graphics::{prelude::*, primitives::Rectangle};
-    ///
-    /// let rect = Rectangle::new(Point::new(10, 20), Size::new(10, 20));
-    /// assert_eq!(rect.is_zero_sized(), false);
-    ///
-    /// let rect = Rectangle::new(Point::new(10, 20), Size::zero());
-    /// assert_eq!(rect.is_zero_sized(), true);
-    /// ```
-    // MSRV: Add const when upgrading to at least 1.46.0
-    pub fn is_zero_sized(&self) -> bool {
-        self.size.height == 0 || self.size.width == 0
     }
 }
 
@@ -462,29 +87,6 @@ impl Transform for Rectangle {
 
         self
     }
-}
-
-/// Anchor point.
-#[derive(Debug, Ord, PartialOrd, Eq, PartialEq, Hash, Copy, Clone)]
-pub enum AnchorPoint {
-    /// Top left.
-    TopLeft,
-    /// Top center.
-    TopCenter,
-    /// Top right.
-    TopRight,
-    /// Center left.
-    CenterLeft,
-    /// Center.
-    Center,
-    /// Center right.
-    CenterRight,
-    /// Bottom left.
-    BottomLeft,
-    /// Bottom center.
-    BottomCenter,
-    /// Bottom right.
-    BottomRight,
 }
 
 #[cfg(test)]
